Analysis of energy consumption
Aim of the project is to analyze the energy consumption data of a house. Measurement of the global consumption of the house is provided by 3 sub meters that measure different zones of the house. The objective is to analyze the data, give consumption advice to the client and to deliver predictions that can generate an economic benefit to the implantation of a smart home.